tests modified, uid added to User model, README modified
