+++ README.md
@@ -1,10 +1,38 @@
-# Доска объявлений
+# Дипломная работа в SkyPro: Доска объявлений
+Бэкенд-часть для сайта объявлений: пользователь может зарегистрироваться, 
+публиковать свои объявления, оставлять отзывы под своими и чужими объявлениями.,
+а также осуществлять поиск по объявлениям.
 
 ## Описание
-Бэкенд-часть для сайта объявлений: авторизация и аутентификация пользователей,
-возможность восстановления пароля через электронную почту, CRUD для объявлений
-на сайте, возможность оставлять отзывы под каждым объявлением, а также осуществлять
-поиск объявлений по названию.
+
+* Django приложение с использованием DRF; 
+* Для сохранения данных спользуется PostgreSQL;
+* Настроена аутентификация и авторизация пользователей с использованием JWT; 
+* Настроен CORS;
+* Имеется список зависимостей;
+* Результат проверки Flake8 равен 100% при исключении миграций;
+* Покрытие тестами 99%;
+* Каждое объявление имеет название, описание, цену и связанные с ним комментарии;
+* Объявления отсортированы по дате добавления;
+* Осуществлена функция поиска по объявлениям;
+* Документация доступна по ссылке http://127.0.0.1:8000/redoc/ и http://127.0.0.1:8000/swagger/;
+* Описаны права доступа:
+  * Просматривать список зарегистрированных пользователей может только администратор;
+  * Просматривать, обновлять и аккаунт пользователя может только администратор и владелец аккаунта;
+  * Удалять аккаунт пользователя может только администратор; 
+  * Создавать объявления и комментарии могут только авторизованные пользователи и админитраторы;
+  * Просматривать список объявлений могут только зарегитрированные пользователи;
+  * Изменять и удалять объявления и комментарии могут только их авторы и админитраторы;
+
+## Технологии
+
+* Python
+* Django, DRF
+* JWT
+* DRF-YASG
+* PostgreSQL
+* Pytest, coverage
+
 
 ## Запуск проекта локально
 1. Клонируйте репозиторий:
@@ -15,25 +43,49 @@ git clone https://github.com/Hallgerd-M/polevaya_message_board
 ```
 pip install -r requirements.txt
 ```
-3. Создайте базу данных и выполните миграции:
+3. Для работы с переменными окружениями создайте файл .env 
+и заполните его согласно файлу .env-emplate
+4. Создайте базу данных и выполните миграции:
 ```
 python manage.py makemigrations
 python manage.py migrate
 ```
-4. Запустите локальный сервер
+4. Для создания администратора нужно запустить команду:
+
+```
+python manage.py csu
+```
+5. Запустите локальный сервер
 ```
 python manage.py runserver
 ```
 5. Через Postman регистрируйте пользователей, добавляйте объявления и отзывы
+6. Для тестирования проекта запустите команду:
 
+```
+python manage.py test
+```
+
+7. Для запуска подсчета покрытия и вывода отчет запустите команды:_
+```
+pytest --nomigrations --cov 
+```
 ## Запуск через Docker
-1. Утановите и откройте программу Docker Desktop: 
+1. Уcтановите и откройте программу Docker Desktop: 
 ```
 https://www.docker.com/products/docker-desktop/
 ```
-2. Запустите в терминале команду:
+2. Для работы с переменными окружениями создайте файл .env 
+и заполните его согласно файлу .env-emplate
+
+3. Для создания образа из Dockerfile и запуска контейнера запустите команду:
 ```
 docker-compose up -d --build
+```
+4. Для тестирования проекта запустите команду:
+
+```
+docker-compose exec app python manage.py test
 ```
 
 ## Документация и структура проекта:
