+++ docker-compose.yml
@@ -8,23 +8,22 @@ services:
     ports:
       - '8000:8000'
     depends_on:
-      db:
-        condition: service_healthy
+      - db
     env_file:
       - ./.env
 
   db:
-    restart: always
     image: postgres:16
+    restart: on-failure
+    ports:
+      - '5432:5432'
     volumes:
-      - postgres_data:/var/lib/postgresql/data/
-    expose:
-      - "5432"
+      - postgres_data:/var/lib/postgresql/data
     env_file:
-      - ./.env
+      - .env
     healthcheck:
-      test: ['CMD-SHELL', 'pg_isready -U postgres']
-      interval: 15s
+      test: [ "CMD-SHELL", "pg_isready -U $POSTGRES_USER" ]
+      interval: 30s
       timeout: 10s
       retries: 5
