+++ pytest.ini
@@ -1,2 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE=config.settings
+python_files = tests.py test_*.py
+addopts = --ignore-glob="*/migrations/*"
