+++ users/models.py
@@ -1,9 +1,12 @@
+import uuid
+
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 
 
 class User(AbstractUser):
     username = None
+    uid = models.UUIDField(unique=True, default=uuid.uuid4, null=True, blank=True)
     email = models.EmailField(unique=True, verbose_name="Email")
     first_name = models.CharField(
         max_length=50,
