--- users/tests.py
@@ -1,27 +1,0 @@
-from django.urls import reverse
-from rest_framework import status
-from rest_framework.test import APITestCase
-
-from users.models import User
-
-
-class UserTestCase(APITestCase):
-
-    def setUp(self):
-        self.user = User.objects.create(email="admin@test.ru")
-        self.client.force_authenticate(user=self.user)
-
-    def test_user_create(self):
-        url = reverse("users:register")
-        data = {"email": "test@test.ru", "password": "test"}
-        response = self.client.post(url, data)
-        print(response.json())
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(User.objects.all().count(), 2)
-
-    def test_user_retrieve(self):
-        url = reverse("users:user_detail", args=(self.user.pk,))
-        response = self.client.get(url)
-        data = response.json()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(data.get("email"), self.user.email)
